refactor: split app entry responsibilities
