+++ cpp/src/app/Application.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <span>
+#include <string_view>
+
+namespace mqb::app {
+
+class Application {
+public:
+    [[nodiscard]] int run(std::span<const std::string_view> arguments);
+};
+
+} // namespace mqb::app
